fix(f2): tetos de materialização, unicidade do caminho e Origem relativa
Adendos 8b e 8c do contrato, mais o polimento apontado na validação.

8b — tetos de materialização. `repeat.n` agora é recusado fora de
[1, LimiteDeClones]. A quantidade declarada fica em float64 até a
resolução: converter para int no schema fazia 1e19 saturar em maxint64
no arm64 (aceito, e um laço de 9,2e18 iterações) e em negativo no amd64
(recusado) — a mesma entrada com dois comportamentos. E o total de
Elementos materializados num Frame é conferido a cada Elemento, contra
LimiteDeElementos, porque Repetições encadeadas por Componentes
multiplicam: oito Componentes com n: 10 cada, dentro do limite de
profundidade, materializavam 10^8 Elementos a partir de menos de 1 KB.
Todos os casos passam a falhar em dezenas de milissegundos, com código 1.

8c — unicidade do <caminho>. As duas regras de segmento podiam colidir
(um Elemento com id: x e um Slot chamado x no mesmo espaço). O caminho
repetido ganha o sufixo ~2, ~3, ... na ordem de pintura. O escopo é o
Frame: dois Frames com os mesmos ids continuam com os mesmos caminhos.

Origem sempre relativa: o diretório do Documento passa a ser
absolutizado, senão um Documento citado em caminho relativo com um
Componente citado em caminho absoluto não tinha denominador comum e o
de= da árvore vazava o caminho absoluto inteiro.

Polimento: Instância e Slot preenchido por Componente compartilham
expande(); materializa, tamanhoNoEixo e segmento tratam o enum por caso
explícito, em vez de um default que desreferenciava no.Caixa contando
com o switch anterior.

Cobertura nova: Repetição escrita dentro do Componente, n: 1, identidade
do Componente pelo caminho absoluto (grafias diferentes do mesmo arquivo
e homônimos em pastas diferentes), precedência do nome do Slot sobre o
id, os dois tetos nas duas bordas, e o de= relativo com use absoluto.

+++ internal/schema/schema.go
@@ -78,8 +78,12 @@ type Disco struct{ X, Y, D float64 }
 
 // Repeticao é a clonagem de um nó ao longo de um eixo, com espaçamento fixo.
 type Repeticao struct {
-	// N é a quantidade de clones, no mínimo 1.
-	N int
+	// N é a quantidade de clones declarada, já arredondada para inteiro. É
+	// float64 porque o teto de clones é da resolução: converter para int
+	// antes dela faria um valor que estoura o int64 saturar num extremo
+	// diferente conforme a plataforma, e a mesma entrada seria aceita numa
+	// e recusada noutra.
+	N float64
 	// Eixo é "x" ou "y".
 	Eixo string
 	// Intervalo é o espaçamento entre clones, em porcentagem do eixo do
